+++ Summator/Summator.cs
@@ -14,11 +14,11 @@
             return sum;
         }
 
-        public static double Average(int[] arr)
+        public static long Average(long[] arr)
         {
-            double sum = 0;          
+            long sum = 0;          
 
-            for (int i = 0; i < arr.Length; i++)
+            for (long i = 0; i < arr.Length; i++)
             {
                 sum += arr[i];
             }
